Generate the FullPath Windows interop with CsWin32
The project already referenced CsWin32, but most of Interop/Windows.cs
was still declaring by hand what the generator can emit from the Win32
metadata: the MAX_PATH, FSCTL_GET_REPARSE_POINT, SYMLINK_FLAG_RELATIVE
and IO_REPARSE_TAG_SYMLINK constants, the WIN32_FIND_DATA and FILE_TIME
structures, a symlink-only view of REPARSE_DATA_BUFFER, and four wrapper
methods whose only job was to cast the BCL enums back to the flags the
native functions expect.

Those are now generated. The call sites call PInvoke.CreateFile,
FindFirstFileEx, DeviceIoControl and GetFinalPathNameByHandle directly
with the real parameter types, and apply PathInternal.EnsureExtendedPrefixIfNeeded
themselves, which was the only thing the wrappers really added.

Two pieces stay hand-written:

- The three error codes. CsWin32 only exposes them as members of
  WIN32_ERROR, and generating that enum adds 217 KB per target framework
  to the assembly for three constants.
- The offset of PathBuffer in REPARSE_DATA_BUFFER. It used to come from
  Marshal.SizeOf of the trimmed-down structure; the full structure is
  larger because it counts the first element of the variable length path
  buffer, so the offset is computed instead of taken from the size.

PathInternal is plain managed path logic rather than interop, so it moves
to its own file and Interop/Windows.cs goes away. The public API does not
change.
