+++ src/Meziantou.Framework.FullPath/NativeMethods.txt
@@ -7,3 +7,9 @@ DeviceIoControl
 FindClose
 FindFirstFileExW
 GetFinalPathNameByHandleW
+MAX_PATH
+FSCTL_GET_REPARSE_POINT
+SYMLINK_FLAG_RELATIVE
+IO_REPARSE_TAG_SYMLINK
+WIN32_FIND_DATAW
+REPARSE_DATA_BUFFER
