+++ src/Meziantou.Framework.FullPath/Symlink.cs
@@ -1,7 +1,11 @@
 using System.Buffers;
 using System.ComponentModel;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.Versioning;
+using Windows.Wdk.Storage.FileSystem;
+using Windows.Win32;
+using Windows.Win32.Storage.FileSystem;
 
 namespace Meziantou.Framework;
 
@@ -134,6 +138,27 @@ internal static partial class Symlink
     [SupportedOSPlatform("windows5.1.2600")]
     private static class WindowsSymlink
     {
+        // CsWin32 only generates these as members of WIN32_ERROR, and that enum carries a few thousand values into the
+        // assembly for the three that are needed here
+        private const int ERROR_SUCCESS = 0x0;
+        private const int ERROR_INSUFFICIENT_BUFFER = 0x7A;
+        private const int ERROR_MORE_DATA = 0xEA;
+
+        /// <summary>
+        /// Byte offset of <c>PathBuffer</c> in <see cref="REPARSE_DATA_BUFFER"/>. The name offsets a symbolic link
+        /// reparse point reports are relative to it. It is not the size of the structure: that also counts the first
+        /// element of the variable length path buffer and the padding the element needs.
+        /// </summary>
+        private static readonly int ReparseDataHeaderLength = GetReparseDataHeaderLength();
+
+        private static int GetReparseDataHeaderLength()
+        {
+            var buffer = default(REPARSE_DATA_BUFFER);
+            return (int)Unsafe.ByteOffset(
+                ref Unsafe.As<REPARSE_DATA_BUFFER, byte>(ref buffer),
+                ref Unsafe.As<char, byte>(ref buffer.SymbolicLinkReparseBuffer.PathBuffer.e0));
+        }
+
         public static bool TryGetSymLinkTarget(string path, [NotNullWhen(true)] out string? target)
         {
             target = null;
@@ -146,16 +171,23 @@ internal static partial class Symlink
             return target is not null;
         }
 
-        internal static bool IsSymbolicLink(string path)
+        internal static unsafe bool IsSymbolicLink(string path)
         {
-            var findData = new Interop.Kernel32.WIN32_FIND_DATA();
-            using var handle = Interop.Kernel32.FindFirstFile(path, ref findData);
+            var findData = default(WIN32_FIND_DATAW);
+
+            // FindExInfoBasic skips the short name, which this code does not read and which costs time to compute
+            using var handle = PInvoke.FindFirstFileEx(
+                PathInternal.EnsureExtendedPrefixIfNeeded(path),
+                FINDEX_INFO_LEVELS.FindExInfoBasic,
+                &findData,
+                FINDEX_SEARCH_OPS.FindExSearchNameMatch,
+                dwAdditionalFlags: default);
             if (!handle.IsInvalid)
             {
                 // dwReserved0 holds the reparse tag, so it must be compared for equality.
                 // A bitwise test would also match other reparse points such as junctions (IO_REPARSE_TAG_MOUNT_POINT).
                 return ((FileAttributes)findData.dwFileAttributes).HasFlag(FileAttributes.ReparsePoint) &&
-                    findData.dwReserved0 == Interop.Kernel32.IO_REPARSE_TAG_SYMLINK;
+                    findData.dwReserved0 == PInvoke.IO_REPARSE_TAG_SYMLINK;
             }
 
             return false;
@@ -164,16 +196,19 @@ internal static partial class Symlink
         // Adapted from dotnet/runtime's reparse point reader, licensed to the .NET Foundation under one or more
         // agreements. The .NET Foundation licenses this file to you under the MIT license.
         // Source: https://github.com/dotnet/runtime/blob/main/src/libraries/System.Private.CoreLib/src/System/IO/FileSystem.Windows.cs
-        internal static string? GetSingleSymbolicLinkTarget(string path)
+        internal static unsafe string? GetSingleSymbolicLinkTarget(string path)
         {
-            using var handle =
-                Interop.Kernel32.CreateFile(path,
-                0,                                                             // No file access required, this avoids file in use
-                FileShare.ReadWrite | FileShare.Delete,                        // Share all access
-                FileMode.Open,
-                Interop.Kernel32.FileOperations.FILE_FLAG_OPEN_REPARSE_POINT | // Open the reparse point, not its target
-                Interop.Kernel32.FileOperations.FILE_FLAG_BACKUP_SEMANTICS);   // Permit opening of directories
-                                                                               // https://docs.microsoft.com/en-us/windows-hardware/drivers/ifs/fsctl-get-reparse-point
+            // https://docs.microsoft.com/en-us/windows-hardware/drivers/ifs/fsctl-get-reparse-point
+            using var handle = PInvoke.CreateFile(
+                PathInternal.EnsureExtendedPrefixIfNeeded(path),
+                dwDesiredAccess: 0,                                                   // No file access required, this avoids file in use
+                FILE_SHARE_MODE.FILE_SHARE_READ | FILE_SHARE_MODE.FILE_SHARE_WRITE |
+                    FILE_SHARE_MODE.FILE_SHARE_DELETE,                                // Share all access
+                lpSecurityAttributes: null,
+                FILE_CREATION_DISPOSITION.OPEN_EXISTING,
+                FILE_FLAGS_AND_ATTRIBUTES.FILE_FLAG_OPEN_REPARSE_POINT |              // Open the reparse point, not its target
+                    FILE_FLAGS_AND_ATTRIBUTES.FILE_FLAG_BACKUP_SEMANTICS,             // Permit opening of directories
+                hTemplateFile: null);
 
             // The link can be deleted, renamed, or have its access denied between the IsSymbolicLink probe and this
             // open. Report that as "no target" so the caller returns false, instead of letting DeviceIoControl fail
@@ -181,47 +216,50 @@ internal static partial class Symlink
             if (handle.IsInvalid)
                 return null;
 
-            var sizeHeader = Marshal.SizeOf<Interop.Kernel32.REPARSE_DATA_BUFFER_SYMLINK>();
-            var bufferSize = sizeHeader + Interop.Kernel32.MAX_PATH;
+            var bufferSize = ReparseDataHeaderLength + (int)PInvoke.MAX_PATH;
 
             while (true)
             {
                 var buffer = ArrayPool<byte>.Shared.Rent(bufferSize);
                 try
                 {
-                    var result = Interop.Kernel32.DeviceIoControl(handle, Interop.Kernel32.FSCTL_GET_REPARSE_POINT, inBuffer: null, cbInBuffer: 0, buffer, (uint)buffer.Length, out var bytesRead, overlapped: IntPtr.Zero) ?
+                    var result = PInvoke.DeviceIoControl(handle, PInvoke.FSCTL_GET_REPARSE_POINT, lpInBuffer: default, buffer, out var bytesRead, lpOverlapped: null) ?
                         0 : Marshal.GetLastWin32Error();
 
-                    if (result is not Interop.Errors.ERROR_SUCCESS and not Interop.Errors.ERROR_INSUFFICIENT_BUFFER and not Interop.Errors.ERROR_MORE_DATA)
+                    if (result is not ERROR_SUCCESS and not ERROR_INSUFFICIENT_BUFFER and not ERROR_MORE_DATA)
                     {
                         throw new Win32Exception(result);
                     }
 
                     ReadOnlySpan<byte> validBuffer = buffer.AsSpan()[..(int)bytesRead];
 
-                    if (!MemoryMarshal.TryRead<Interop.Kernel32.REPARSE_DATA_BUFFER_SYMLINK>(validBuffer, out var header))
+                    if (validBuffer.Length < ReparseDataHeaderLength)
                     {
-                        if (result == Interop.Errors.ERROR_SUCCESS)
+                        if (result == ERROR_SUCCESS)
                         {
                             // didn't read enough for header
                             throw new InvalidDataException("FSCTL_GET_REPARSE_POINT did not return sufficient data");
                         }
 
                         // can't read header, guess at buffer length
-                        bufferSize = checked(buffer.Length + Interop.Kernel32.MAX_PATH);
+                        bufferSize = checked(buffer.Length + (int)PInvoke.MAX_PATH);
                         continue;
                     }
 
+                    // The rented array is always at least as long as the fixed part of the structure, so the header can be
+                    // read in place. Only the fields before PathBuffer are touched until the payload length is validated.
+                    ref var header = ref Unsafe.As<byte, REPARSE_DATA_BUFFER>(ref MemoryMarshal.GetArrayDataReference(buffer)).SymbolicLinkReparseBuffer;
+
                     // we only care about SubstituteName.
                     // Per https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-fscc/b41f1cbf-10df-4a47-98d4-1c52a833d913 print name is only valid for displaying to the user
-                    bufferSize = sizeHeader + header.SubstituteNameOffset + header.SubstituteNameLength;
-                    // bufferSize = sizeHeader + Math.Max(header.SubstituteNameOffset + header.SubstituteNameLength, header.PrintNameOffset + header.PrintNameLength);
+                    bufferSize = ReparseDataHeaderLength + header.SubstituteNameOffset + header.SubstituteNameLength;
+                    // bufferSize = ReparseDataHeaderLength + Math.Max(header.SubstituteNameOffset + header.SubstituteNameLength, header.PrintNameOffset + header.PrintNameLength);
 
                     if (bytesRead >= bufferSize)
                     {
                         // got entire payload with valid header.
-                        var target = Encoding.Unicode.GetString(validBuffer.Slice(sizeHeader + header.SubstituteNameOffset, header.SubstituteNameLength));
-                        if ((header.Flags & Interop.Kernel32.SYMLINK_FLAG_RELATIVE) != 0)
+                        var target = Encoding.Unicode.GetString(validBuffer.Slice(ReparseDataHeaderLength + header.SubstituteNameOffset, header.SubstituteNameLength));
+                        if ((header.Flags & Windows.Wdk.PInvoke.SYMLINK_FLAG_RELATIVE) != 0)
                         {
                             // The device prefix is taken off first, so the relative target is combined with an ordinary path
                             var isExtended = PathInternal.IsExtended(path);
